projects/ad4130_iio/scripts/requirements.txt
numpy==1.22.4
pyadi-iio==0.0.14
pylibiio==0.23.1
pynput==1.7.6
pyserial==3.5